feat: move function to utils
