+++ src/utils/index.ts
@@ -1,2 +1,3 @@
 export * from "./ajv";
 export * from "./address";
+export * from "./github";
